services: don't kill all jobs every time
Have each job "adopt" the still-applicable jobs from the previous
services version.

The prior version did:

    let services = chan.recv();
    loop {
        started = start(services);
        services = chan.recv();
        kill(started);
    }

The new version is a bit simpler maybe? It does this:

chan
    .iter()
    .fold(|previous_services, new_services| {
       start(new_services)
    })
    .for_each(|_| {});

where the return of `start` has a Drop implementation to call
`kill()`.

This use of `fold` is nice because it allows us to "adopt" processes
from prior iterations in to the new iteration.